[SIX-186] feat : 특정 유저 리뷰 조회 컨트롤러 패키지 변경
